Added Elemental Weaknesses and Resistances to Battle System
Using the WeaponElement and UnitElement enums from the Unit and Weapon scriptable objects, a CalculateDamage function was added to the battle system to determine how much damage an enemy would take based on weapon typing and enemy typing. Also added if statement to button to only work if durability is greater than 0.