CSV Exports to Scriptable Objects
All the spreadsheet work can now be added as SOs with the press of a button. On the top of Unity, there should be a Utility dropdown and an option to import enemies or weapons. Just needs a CSV file with the exact same names as the ones already there.